+++ crates/rtlola-hir/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add '->' syntactic sugar
 - Added `window_num_buckets` method to `MemBoundTrait`.
 - Added `sliding_window_bucket_size` method to `MemBoundTrait`.
+- Added support for multiple eval clauses.
 
 ### Changed
 - Applying the `abs()` function to unsigned values is now rejected by the typechecker.